changed flight search to a component, also made the functionality to remove the return date input field if the flight type is "one way"
